Update PK_ANIMAL_RESOURCE.md
As it stands, there's a discrepancy on whether we're using the `__init__` file, and this appears to the the first time in the book an `animals` file was mentioned. If we're still using `__init__` files for package management, we should have the students add the function to the init, and then import it from the package.